[CI skip] Refactoring and code cleanup
