+++ src/main/java/io/github/thebusybiscuit/slimefun4/api/MinecraftVersion.java
@@ -1,5 +1,7 @@
 package io.github.thebusybiscuit.slimefun4.api;
 
+import javax.annotation.Nonnull;
+
 import org.apache.commons.lang.Validate;
 import org.bukkit.Server;
 
@@ -59,7 +61,7 @@ public enum MinecraftVersion {
      * @param name
      *            The display name of this {@link MinecraftVersion}
      */
-    MinecraftVersion(String name) {
+    MinecraftVersion(@Nonnull String name) {
         this(name, false);
     }
 
@@ -73,7 +75,7 @@ public enum MinecraftVersion {
      * @param virtual
      *            Whether this {@link MinecraftVersion} is virtual
      */
-    MinecraftVersion(String name, boolean virtual) {
+    MinecraftVersion(@Nonnull String name, boolean virtual) {
         this.name = name;
         this.virtual = virtual;
         this.prefix = name().replace("MINECRAFT_", "v") + '_';
@@ -84,6 +86,7 @@ public enum MinecraftVersion {
      * 
      * @return The name of this {@link MinecraftVersion}
      */
+    @Nonnull
     public String getName() {
         return name;
     }
@@ -110,8 +113,9 @@ public enum MinecraftVersion {
      * 
      * @return Whether the version matches with this one
      */
-    public boolean matches(String version) {
+    public boolean matches(@Nonnull String version) {
         Validate.notNull(version, "The input version must not be null!");
+
         return version.startsWith(prefix);
     }
 
@@ -126,7 +130,7 @@ public enum MinecraftVersion {
      * 
      * @return Whether this {@link MinecraftVersion} is newer or equal to the given {@link MinecraftVersion}
      */
-    public boolean isAtLeast(MinecraftVersion version) {
+    public boolean isAtLeast(@Nonnull MinecraftVersion version) {
         Validate.notNull(version, "A Minecraft version cannot be null!");
 
         if (this == UNKNOWN) {
@@ -146,7 +150,7 @@ public enum MinecraftVersion {
      * 
      * @return Whether this {@link MinecraftVersion} is older than the given one
      */
-    public boolean isBefore(MinecraftVersion version) {
+    public boolean isBefore(@Nonnull MinecraftVersion version) {
         Validate.notNull(version, "A Minecraft version cannot be null!");
 
         if (this == UNKNOWN) {
