+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/attributes/PiglinBarterDrop.java
@@ -3,7 +3,7 @@ package io.github.thebusybiscuit.slimefun4.core.attributes;
 import org.bukkit.entity.Piglin;
 import org.bukkit.event.entity.EntityDropItemEvent;
 
-import io.github.thebusybiscuit.slimefun4.implementation.listeners.PiglinListener;
+import io.github.thebusybiscuit.slimefun4.implementation.listeners.entity.PiglinListener;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
