+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/networks/cargo/CargoUtils.java
@@ -30,7 +30,7 @@ final class CargoUtils {
     /**
      * These are the slots where our filter items sit.
      */
-    static final int[] FILTER_SLOTS = { 19, 20, 21, 28, 29, 30, 37, 38, 39 };
+    private static final int[] FILTER_SLOTS = { 19, 20, 21, 28, 29, 30, 37, 38, 39 };
 
     private CargoUtils() {}
 
@@ -371,10 +371,23 @@ final class CargoUtils {
     /**
      * Get the whitelist/blacklist slots in a Cargo Input Node. If you wish to access the items
      * in the cargo (without hardcoding the slots in case of change) then you can use this method.
+     * 
+     * @deprecated Renamed to {@link #getFilteringSlots()}
      *
      * @return The slot indexes for the whitelist/blacklist section.
      */
+    @Deprecated
     public static int[] getWhitelistBlacklistSlots() {
+        return FILTER_SLOTS;
+    }
+
+    /**
+     * Gets the {@link ItemFilter} slots for a Cargo Node. If you wish to access the items
+     * in the cargo (without hardcoding the slots in case of change) then you can use this method.
+     *
+     * @return The slots where the {@link ItemFilter} section for a cargo node sits
+     */
+    public static int[] getFilteringSlots() {
         return FILTER_SLOTS;
     }
 }
