+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/networks/cargo/ItemFilter.java
@@ -92,7 +92,7 @@ class ItemFilter implements Predicate<ItemStack> {
             this.checkLore = Objects.equals(blockData.getString("filter-lore"), "true");
             this.rejectOnMatch = !Objects.equals(blockData.getString("filter-type"), "whitelist");
 
-            for (int slot : CargoUtils.FILTER_SLOTS) {
+            for (int slot : CargoUtils.getFilteringSlots()) {
                 ItemStack stack = menu.getItemInSlot(slot);
 
                 if (stack != null && stack.getType() != Material.AIR) {
