+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/listeners/entity/BeeListener.java
@@ -1,4 +1,4 @@
-package io.github.thebusybiscuit.slimefun4.implementation.listeners;
+package io.github.thebusybiscuit.slimefun4.implementation.listeners.entity;
 
 import java.util.Optional;
