+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/listeners/entity/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package contains any {@link org.bukkit.event.Listener} from Slimefun which is related to an
+ * {@link org.bukkit.entity.Entity}.
+ */
+package io.github.thebusybiscuit.slimefun4.implementation.listeners.entity;
